moving all the device id auth stuff to IAuthApi ( because we want it to be UNITY SDK only thing, SystemInfo class does  not work in the common namespace because it is used both by unity client side and microservices )
