+++ client/Packages/com.beamable/Common/Runtime/Api/Auth/IAuthApi.cs
@@ -6,15 +6,15 @@ namespace Beamable.Common.Api.Auth
       Promise<User> GetUser(TokenResponse token);
       Promise<bool> IsEmailAvailable(string email);
       Promise<bool> IsThirdPartyAvailable(AuthThirdParty thirdParty, string token);
-      Promise<bool> IsThisDeviceIdAvailable();
+
       Promise<TokenResponse> CreateUser();
       Promise<TokenResponse> LoginRefreshToken(string refreshToken);
       Promise<TokenResponse> Login(string username, string password, bool mergeGamerTagToAccount = true, bool customerScoped=false);
       Promise<TokenResponse> LoginThirdParty(AuthThirdParty thirdParty, string thirdPartyToken, bool includeAuthHeader = true);
-      Promise<TokenResponse> LoginDeviceId();
+
       Promise<User> RegisterDBCredentials(string email, string password);
       Promise<User> RegisterThirdPartyCredentials(AuthThirdParty thirdParty, string accessToken);
-      Promise<User> RegisterDeviceId();
+
       Promise<EmptyResponse> IssueEmailUpdate(string newEmail);
       Promise<EmptyResponse> ConfirmEmailUpdate(string code, string password);
       Promise<EmptyResponse> IssuePasswordUpdate(string email);
