+++ client/Packages/com.beamable/Runtime/Core/Platform/SDK/Auth/AuthService.cs
@@ -1,5 +1,8 @@
+using Beamable.Common;
 using Beamable.Common.Api;
 using Beamable.Common.Api.Auth;
+using System;
+using UnityEngine;
 
 namespace Beamable.Api.Auth
 {
@@ -17,7 +20,10 @@ namespace Beamable.Api.Auth
    /// </summary>
    public interface IAuthService : IAuthApi
    {
-   }
+        Promise<bool> IsThisDeviceIdAvailable();
+        Promise<User> RegisterDeviceId();
+        Promise<TokenResponse> LoginDeviceId();
+    }
 
    /// <summary>
    /// This type defines the %Client main entry point for the %Auth feature.
@@ -37,5 +43,43 @@ namespace Beamable.Api.Auth
       {
       }
 
-   }
+        public Promise<bool> IsThisDeviceIdAvailable()
+        {
+            var encodedDeviceId = Requester.EscapeURL(SystemInfo.deviceUniqueIdentifier);
+            return Requester.Request<AvailabilityResponse>(Method.GET, $"{ACCOUNT_URL}/available/device-id?deviceId={encodedDeviceId}", null, false)
+                .Map(resp => resp.available);
+        }
+
+        public Promise<TokenResponse> LoginDeviceId()
+        {
+            var req = new LoginDeviceIdRequest
+            {
+                grant_type = "device",
+                device_id = SystemInfo.deviceUniqueIdentifier
+            };
+            return Requester.Request<TokenResponse>(Method.POST, TOKEN_URL, req);
+        }
+
+        public class LoginDeviceIdRequest
+        {
+            public string grant_type;
+            public string device_id;
+        }
+
+        public Promise<User> RegisterDeviceId()
+        {
+            var req = new RegisterDeviceIdRequest
+            {
+                deviceId = SystemInfo.deviceUniqueIdentifier
+            };
+            return Requester.Request<User>(Method.PUT, $"{ACCOUNT_URL}/me", req);
+        }
+
+        [Serializable]
+        private class RegisterDeviceIdRequest
+        {
+            public string deviceId;
+        }
+
+    }
 }
